refactor the post search and pagination dto's
