--- src/api/v1/dtos/searchDto.js
@@ -1,90 +1,0 @@
-import Joi from 'joi'
-import { POST_SORT_OPTIONS } from '../../../config/dbConstants/entityTypeIds.js'
-import { queryParamsValidatorBuilder } from '../../helpers/index.js'
-import { uidSchema } from '../middleware/validateUids.js'
-
-// Schemas
-const querySchema = Joi.string().trim().required().messages({
-  'string.empty': "El valor de 'q' no puede estar vacío",
-  'any.required': "Debes ingresar un título en el parámetro de consulta 'q'"
-})
-
-const orderSchema = Joi.string().valid('asc', 'desc').insensitive().messages({
-  'any.only': "El valor de 'order' solo puede ser 'asc' o 'desc'"
-})
-
-const postSortSchema = Joi.string().valid('votes', 'date').insensitive().messages({
-  'any.only': "El valor de sort solo puede ser 'votes' o 'date'",
-  'string.empty': 'El valor de sort no puede estar vacío'
-})
-
-const pageSchema = Joi.number().integer().optional().messages({
-  'number.base': "El valor de 'page' debe ser un número",
-  'number:integer': "El valor de 'page' debe ser un número entero"
-})
-
-const limitSchema = Joi.number().integer().optional().messages({
-  'number.base': "El valor de 'limit' debe ser un número",
-  'number:integer': "El valor de 'limit' debe ser un número entero"
-})
-
-const postSearchSchema = Joi.object({
-  q: querySchema,
-  sort: postSortSchema.optional(),
-  order: Joi.when('sort', {
-    is: Joi.exist(),
-    then: orderSchema.optional(),
-    otherwise: Joi.forbidden().messages({
-      'any.unknown': "Solo se puede proporcionar un valor de 'order' junto con uno de 'sort'"
-    })
-  }).messages({
-    'string.empty': "El valor de 'order' no puede estar vacío"
-  }),
-  category: uidSchema.label('El valor de category'),
-  page: pageSchema,
-  limit: limitSchema
-}).options({ abortEarly: false }).messages({
-  'object.unknown': "Las únicas queries permitidas son: 'q', 'sort', 'order', 'category', 'page' y 'limit'"
-})
-
-const postPaginationSchema = Joi.object({
-  sort: postSortSchema.optional(),
-  order: Joi.when('sort', {
-    is: Joi.exist(),
-    then: orderSchema.optional(),
-    otherwise: Joi.forbidden().messages({
-      'any.unknown': "Solo se puede proporcionar un valor de 'order' junto con uno de 'sort'"
-    })
-  }).messages({
-    'string.empty': "El valor de 'order' no puede estar vacío"
-  }),
-  category: uidSchema.label('El valor de category'),
-  page: pageSchema,
-  limit: limitSchema
-}).options({ abortEarly: false }).messages({
-  'object.unknown': "Las únicas queries permitidas son: 'sort', 'order', 'category', 'page' y 'limit'"
-})
-
-// DTO functions
-const validatePostSearch = async ({ query }) => await postSearchSchema.validateAsync(query)
-const validatePostPagination = async ({ query }) => await postPaginationSchema.validateAsync(query)
-
-const transformPostPagination = ({ query }) => {
-  const { sort, page, limit } = query
-  query.sort = sort ? POST_SORT_OPTIONS[sort] : POST_SORT_OPTIONS.votes
-  query.order = sort && query.order ? query.order : 'desc'
-  query.page = page ? +page : 1
-  query.limit = limit ? +limit : 20
-  return query
-}
-
-const transformPostSearch = ({ query }) => {
-  query.q = `%${query.q}%`
-  return transformPostPagination({ query })
-}
-
-// DTOs
-const postSearchDto = queryParamsValidatorBuilder(validatePostSearch, transformPostSearch)
-const postPaginationDto = queryParamsValidatorBuilder(validatePostPagination, transformPostPagination)
-
-export { postSearchDto, postPaginationDto }
